refactor: /post, /post/{id}}/message 페이지 이름 변경
