+++ src/routes/Router.jsx
@@ -1,15 +1,16 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Post from "../pages/Post";
-import List from "../pages/list";
 import Main from "../pages/main";
-import PostMessage from "../pages/PostMessage";
+import PostCreate from "../pages/PostCreate";
+import MessageWrite from "../pages/MessageWrite";
+import List from "../pages/list";
 
 function Router() {
   return (
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Main />} />
-        <Route path="/post" element={<Post />} />
+        <Route path="/post" element={<PostCreate />} />
+        <Route path="/Post/{id}/Message" element={<MessageWrite />} />
         <Route path="/list" element={<List />} />
       </Routes>
     </BrowserRouter>
